requirements.txt
asgiref==3.2.10
Django==3.1.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.1
flake8==3.8.4
greenlet==0.4.17
mccabe==0.6.1
msgpack==1.0.0
neovim==0.3.1
pycodestyle==2.6.0
pyflakes==2.2.0
pynvim==0.4.2
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
